Fixing side set highlighting
Fixing a long-standing problem with side set high-lighting. We were
scaling down the volumetric blocks, so that the side set could be
drawn "above" it.  This meant that we were reporting incorrect domain
dimensions. This also meant that on large domains the side sets were
drawn slightly off, since the scaling error got big enough.